refactor : List 로 Lotto들을 관리하는 방식에서 Lottos 객체에서 관리하도록 변경
